Refactor the Pwndoc's backend Docker image
The new docker image changes the way configuration options are passed
to the application. Instead of using config.json all the required
information are passed as environment variables.

Another notable change in the new image is related to the certificates.
Instead of using the hardcoded certificate, the application will generate
a CA and self-signed certificates for backend and frontend at runtime.